+++ src/spot.js
@@ -1,12 +1,12 @@
 const APIBase = require('./APIBase')
 const {
   Blvt, Bswap, SubAccount, Market, Trade, Futures, Fiat,
-  Wallet, Margin, Mining, Savings, Stream, Websocket
+  Wallet, Margin, Mining, Savings, Stream, Websocket, C2C
 } = require('./modules')
 const { flowRight } = require('./helpers/utils')
 
 class Spot extends flowRight(Blvt, Bswap, SubAccount, Websocket, Stream,
-  Savings, Margin, Mining, Wallet, Market, Trade, Futures, Fiat)(APIBase) {
+  Savings, Margin, Mining, Wallet, Market, Trade, Futures, Fiat, C2C)(APIBase) {
   constructor (apiKey = '', apiSecret = '', options = {}) {
     options.baseURL = options.baseURL || 'https://api.binance.com'
     super({