+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
-##  v1.1.0
+## v1.2.0 - 2021-08-20
+### Added
+- `GET /sapi/v1/c2c/orderMatch/listUserOrderHistory` to query user C2C trade history
+- `GET api/v3/myTrades` has a new optional field orderId
+
+## v1.1.0 - 2021-08-12
 ### Added
 - `GET /sapi/v1/fiat/orders` to query user fiat deposit and withdraw history
 - `GET /sapi/v1/fiat/payments` to query user fiat payments history
 
 
-## v1.0.0
+## v1.0.0 - 2021-07-27
 ### Added
 - First release, please find details from `README.md`